Move copy_directory to helpers
