+++ test/helpers/copy_directory.go
@@ -1,4 +1,4 @@
-package runtime
+package helpers
 
 import (
 	"fmt"
